Teachers: Prevent columns from overlapping 'create teacher' button.
When hovering over the 'Create Teacher' button, a user will
find that they may only click on the left half of the button.
This is terrible, so widen the middle column containing the form,
and narrow the outer columns, so the layout doesn't interfere
with the active area of the button. OMG\!