RegistrationsController: Conditionally add existing activity to list.
Retrieving the available activities might not include the one the
student is registered for, in the case that it is at capacity. But,
we want that to appear in the list, as the student is registered
for it during the change. However, just explicitly adding it to the
list will result in a dup activity in the list, when the available
activities includes the one the student is registered for. So, only
add it to the list if it's not already there. (Shameless - refactor\!)

References #94.