Nav: Add last name A to main studentss nav item.
When viewing the students list, the 'A' block in the paginator isn't
highlighted. This is driven by a querystring.

Add the querystring to the main 'Students' navbar link, so that when
a user initially views the list of students, which shows only the 'A'
students, the paginator is appropriately highlighted.