Routes: Set new root url, and change redirects.
The root_url is activities#index but we want to use the new static home
page as the de facto root url. However, we are redirecting to root_url
in many places throughout the codebase. This root url was convenient
when we had no 'home' page.

Set the root route to the new static home page.
Update all redirects to root_url that are expecting activities#index.
Update the tests to reflect this change of behavior.

References #142.