Migrations: Fix null constraint.
The school association migrations set a null constraint to be false, which
allows for null values. This was an accident - we do not want to allow
null values on these school_id columns.

Fix the migrations. This does not require a new migration, as we have not
applied these migrations anywhere else, such as staging/production, other
devenvs, etc.

References #143.