AuthenticatedController: Introduce and subclass.
With acts_as_tenant, we can encounter problems with static pages
and devise controllers. This is because setting the tenant assumes
that the user is signed in and relies on current_user to obtain the
school. But the static pages and devise controllers don't have a
signed in user, so current_user is nil.

Introduce an AuthenticatedController that the core app controllers
inherit from, and that contains all the authy-rule stuff and the
setting of the tenant. Let ApplicationController be pretty dumb,
and let the static pages controller and devise controllers inherit
from it.

Change the rules of the authy before filters, as we'll assume they
will never be fired in a devise controller and don't need opted out of.