Attendance: Test registration controller mark_attendance multi-tenancy.
The controller should not be able to find the other school registration,
and should send a redirect. Note that this redirect goes to the student#show
action, which would also be out of scope for the current tenant. The
StudentsController would send another redirect. But, we do not test for that
here, we'll test for it in the StudentsController test.

References #157.