Activity: Set tenant in test.
I introduced a test failure in c434a89c2d4d4a373d7759fb70de6214fa40f23e, which
added a new activity for a second school in the fixtures. This causes a failure
in the model test, since no tenant is set, resulting in all activities regardless
of school to be retrieved.

Scope all queries in the Activity model test to use the first school as the
tenant.