User: Remove custom teacher multi-tenant validation.
In 0b4fcbc1 I created a my own validation method to ensure that a user's teacher
is from the same school as the user. But, acts_as_tenant enforces this for us.
It wasn't doing it due to a now-known issue that has been addressed (reordering).

Remove the custom validator, keep the test, but also put the test in the
unit test instead of the system test file (it was in the wrong place).

References #161