Registration: Replace custom validator with acts_as_tenant behavior.
I created a custom validator because of a bug regarding ordering of the
acts_as_tenant declaration, which resulted in out-of-scope association
validation errors to not surface.

Since the order of the acts_as_tenant declaration is now fixed, there is
no longer a need to use a custom validator. Remove it, and add tests
that assert the reliability of the acts_as_tenant validations.