User: redirect when student user not found.
When signing in for the first time, a staff member has a
role of 'student' and sees the welcome screen. They are
prompted to select their flextime teacher. But before
anything else happens the admin changes the staff member's
role from 'student' to 'staff'. If the user then tries
to click on anything, or update the flextime teacher as
prompted, an exception is raised, because the controller
can't find a user with the student role any longer.

Rescue the exception and redirect the newly-set staff
user to main schedule screen.

References #178.