User: Use placeholder string for unknown last name.
Commit c7837096ff3832e9c4b68cdf2fc9e765aca331b4 introduced a bug:
it was using an empty string when the last name wasn't in the auth
payload, and last names can't be blank.

This causes first_or_create to return an unsaved user (validations fail)
which then results in the auth controller to send a redirect and deny
the sign in attempt.

Replace the empty string with 'Unknown'.

References #182