Registration: Test staff editing past registrations.
A pending, empty test exists for changing past registrations, and it has
now surfaced as a potential bug / untested feature.

We do indeed wish to enable admins and staff to change past registrations,
because unassigned students may need to be assigned an activity by a
staff member after the cutoff time.

References #181.