Registration: Specify creator during update.
When a student self-registers for an activity, they are the 'creator' of
that registration. Unfortunately, there is logic in Registration that
prevents a registration being for a restricted activity when the creator
of the registration is a student. This prevents students from self-registering
for restricted activities at the model layer.

Unfortunately, this also prevents staff from modifying an existing registration
that the student created. Even though the current user is staff, and they should
be able to update the registration to use a restricted activity, they cannot.
This is because the controller is updating the existing Registration object,
leaving the creator as the student, and failing to save.

Re-set the Registration creator with the current_user during an update,
enabling staff to change an existing student registration to a restricted
activity. Add a flash to the update failure too, to facilitate future
conversations if a problem in this area of the controller arises again.

References #183.